Added the inter frame delay to the send method
The inter frame delay were added to the send method to ensure that the receiving device detects the end of a frame. The internal standard delay function can now be replaced by a custom delay function. Lower baud rates resulting in higher Modbus delays and if an RTOS is used a specific delay can be implemented to delay a caling task under appropriate system requirements.