Reworked the begin() implementation
Outsourced most of the begin() implementation to the SerialModbusBase class and removed the default values - so a pointer to a Serial object is now a mandatory parameter.
begin() now also have a return value and the ICT (Inter Character Timeout) and IFD (Inter Frame Delay) are now always calculated directly.