Improved request and request map handling
It is now possible to set single specific requests besides and/or while using the request map.
E.g., we can have a map with cyclic request, and also use event driven single request.
Further it is also possible to exchange the complete request map while running (which was not possible before, because the map index handling was faulty).
NOTE: Previously forgotten remainings of getReplyData() were also removed.