Added new compatibility management
The compatibility is now managed with the defines provided by the Arduino environmentt, regarding specific Arduino boards and their architecture. If a selected board and/or the architecture have not yet been tested, the IDE/gcc throws a warning and assumed default values for the serial port will be used.